+++ src/app/mega-menu-header/components/footer/index.tsx
@@ -1,0 +1,11 @@
+import { NavAndSearch } from '@/app/mega-menu-header/components/footer/nav-and-search';
+import { Options } from '@/app/mega-menu-header/components/footer/options';
+
+export const Footer = () => {
+  return (
+    <footer>
+      <NavAndSearch />
+      <Options />
+    </footer>
+  );
+};
