+++ src/app/mega-menu-header/components/header/mega-menu/menu3/category-list/index.tsx
@@ -1,6 +1,6 @@
-import { List } from '@/app/mega-menu-header/header/mega-menu/list';
-import { categories } from '@/app/mega-menu-header/header/mega-menu/menu3/data';
-import styles from '@/app/mega-menu-header/header/mega-menu/menu3/styles.module.css';
+import { List } from '@/app/mega-menu-header/components/header/mega-menu/list';
+import { categories } from '@/app/mega-menu-header/components/header/mega-menu/menu3/data';
+import styles from '@/app/mega-menu-header/components/header/mega-menu/menu3/styles.module.css';
 
 export const CategoryList = () => (
   <div className={`${styles.contents} ${styles.contentWithBorder}`}>
