+++ src/app/mega-menu-header/components/header/mega-menu/menu4/terms-list/index.tsx
@@ -1,6 +1,6 @@
-import { List } from '@/app/mega-menu-header/header/mega-menu/list';
-import { terms1, terms2 } from '@/app/mega-menu-header/header/mega-menu/menu4/data';
-import styles from '@/app/mega-menu-header/header/mega-menu/menu4/styles.module.css';
+import { List } from '@/app/mega-menu-header/components/header/mega-menu/list';
+import { terms1, terms2 } from '@/app/mega-menu-header/components/header/mega-menu/menu4/data';
+import styles from '@/app/mega-menu-header/components/header/mega-menu/menu4/styles.module.css';
 
 export const TermsList = () => (
   <div className={styles.content}>
