--- src/app/mega-menu-header/footer/index.tsx
@@ -1,11 +1,0 @@
-import { NavAndSearch } from '@/app/mega-menu-header/footer/nav-and-search';
-import { Options } from '@/app/mega-menu-header/footer/options';
-
-export const Footer = () => {
-  return (
-    <footer>
-      <NavAndSearch />
-      <Options />
-    </footer>
-  );
-};
